# [Codegen][GPU] Adding scheduling barrier between compute and write stage in prefetcher