--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/ROCDLPrefetching.cpp
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/ROCDLPrefetching.cpp
@@ -7,6 +7,7 @@
 #include "iree/compiler/Codegen/LLVMGPU/Passes.h"
 #include "iree/compiler/Codegen/LLVMGPU/Utils/LLVMGPUUtils.h"
 #include "llvm/ADT/SmallVector.h"
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUDialect.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Dialect/SCF/Transforms/Transforms.h"
 #include "mlir/IR/PatternMatch.h"