Motivation: I and @nirvedhmeshram observed a fair amount of misplacement of `s_waitcnt` instructions from the amdgpu backend. The backend tends to place the `s_waitcnt` overly early, sometimes in between MFMA instructions. This has caused irrelevant stalls between compute and load that should have overlap and can potentially hide the stalls.

This PR adds the scheduling barrier in between compute and write stage such that the backend compiler no longer can freely place the `s_waitcnt` earlier, in compute stage. It is forced to place the wait as early as the start of the write stage. 

I've verified through thread-trace that the scheduling barrier does what is intended and the `s_waitcnt` placement isn't interleaved with MFMA instructions any longer.

Upon looking at CI result, this seems to provide a bit of overall lift to all model level tests. Compared target (tip of main at): https://github.com/iree-org/iree/actions/runs/15767966101/job/44447951812



| Benchmark | Before    |   After  |
| --- | --- | --- |
| llama 8b_f16_decode | 10.020865534565278 ms | 9.531815848439127 ms |
| llama 8b_f16_prefill | 39.90881894197729 ms | 39.257620445763074 ms |
| sdxl clip | 8.061422783367593 ms | 7.982292237945579 ms |
| sdxl e2e | 286.39431609772146 ms | 285.5649597477168 ms |
| sdxl punet_int8_fp16 | 40.25666919153404 ms | 39.64519846356578 ms |
| sdxl punet_int8_fp8 | 42.10774295360727 ms | 40.13837954467711 ms |
| sdxl unet_fp16 | 68.40701173059642 ms | 68.04983000271024 ms |
| sdxl vae | 68.57368536293507 ms | 67.36758409999312 ms |