@jerryyin 2025-06-20
@kuhar Good point. I think the scheduling barrier has existed for a decent amount of time that the ROCm side targets we cared about (CDNA and RDNA included) all have it. 

Regarding on sched_barrier placement, I think it will be hard to reliably do it when it has lowered towards rocDL dialect level. We'll have to re-parse the entire IR to be able to recognize what section of IR belong to what stage, effectively duplicating this pass in a much later stage.

I'm not sure I fully understand the part of the suggestion related with the generic barrier op part. Could you elaborate a bit on that? The most intuitive thought I have right now is basically to add a condition check on the `TargetAttr` and only add the scheduling barrier when it is a amdgpu target. Does that sounds good enough to you?