@jerryyin 2025-06-20
> gfx10+ uses a more fine-grained waitcnt

Good to know and thanks for the slides! Although I doubt how much difference that will make because the bottleneck of conv/gemm are usually in buffer_loads so not sure separating the load and store count can be as useful. At this location in particular, stores hasn't been issued yet, and will only be issued at write stage. 

> One alternative is to add a similar op in IREE/mlir that later lowers to the rocdl scheduling barrier op.

IIRC, I think one distinction between sched_barrier and gpu.barrier is that sched_barrier only exists in amdgpu backend. So I'm really hesitant about adding an op that will apply to only a particular target but yield no-op in other targets. 

Having went through this useful discussion, I tend to apply the original proposal of using a straightforward conditional check, and maybe a TODO for us to come back and generalize or delete as needed. Let me know if you have any objections to this. Thanks!