@jerryyin 2025-06-20
I can see why you think that is the case. BTW, that pass is also from me. I can open a new ticket to address that and we can continue to discuss separately so I don't end up to be the guy that disrupt iree coding standard :-p.

For the context of this PR, what would you suggest as a non-layer violating way to implement this? Do you agree about my concerns about building a target independent op that's unique in amdgpu target? We can host a separate meeting if that turns out to be necessary.