@kuhar 2025-06-20
> I can open a new ticket to address that and we can continue to discuss separately so I don't end up to be the guy that disrupt iree coding standard :-p.

SG.

> what would you suggest as a non-layer violating way to implement this? Do you agree with my concerns about building a target independent top-level op that's actually unique in amdgpu target?

I left a couple suggestions at the end of this comment: https://github.com/iree-org/iree/pull/21151#pullrequestreview-2946664801 . IMO if having a generic op that may end up being a nop feels weird, emitting a target-specific op in this pass is 10x worse -- we are adding very low-level hints relatively high up in the compilation pipeline. To me having a target-independent op is lesser evil: it fixes layering violations at the level of MLIR dialects / build system, and allows you to delay deciding what to do with it when you have target information available and useable. Alternatively, make this a rocm pass if you don't care about nvidia anyway.